// src/models/enums.ts
export enum InboundEventType {
  MESSAGE = 'message',
  STATUS = 'status',
}

export enum ChannelType {
  RCS = 'rcs',
}

export enum BrokerType {
  PONTAL_TECH = 'pontal-tech',
}

export enum MessageDirection {
  INBOUND = 'inbound',
  OUTBOUND = 'outbound',
}

export enum MessageStatus {
  QUEUED = 'queued',
  SENT = 'sent',
  DELIVERED = 'delivered',
  READ = 'read',
  ERROR = 'error',
}